requirements.txt
pandas>=2.1.3
tqdm>=4.66.1
numpy>=1.26.2
torch>=2.4.1
transformers>=4.45.2
tiktoken>=0.5.2
langchain>=0.0.345
pygooglenews>=0.1.2
opendartreader>=0.2.3
urllib3>=2.1.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.2